Remove redundant LevelTests
